requirements.txt
altgraph==0.17.2
asgiref==3.5.0
astroid==2.9.3
auto-py-to-exe==2.15.0
autopep8==1.7.0
bandit==1.7.4
black==22.8.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.13.0
django-environ==0.8.1
django-jazzmin==2.6.0
django-templated-mail==1.1.1
django-use-email-as-username==1.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
Eel==0.12.4
filelock==3.4.2
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
gitdb==4.0.9
GitPython==3.1.29
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
mailjet-rest==1.3.4
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.941
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==21.3
pathspec==0.10.1
pbr==5.11.0
pefile==2021.9.3
Pillow==9.0.1
pipenv==2021.11.23
platformdirs==2.4.1
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
PyJWT==2.6.0
pylint==2.12.2
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2021.3
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
smmap==5.0.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stevedore==4.1.0
stripe==4.2.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
whichcraft==0.6.1
whitenoise==6.0.0
wrapt==1.13.3
zope.event==4.5.0
zope.interface==5.4.0
